pyproject.toml
[project]
name = "trend-korea-2026n"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ddgs>=7.2.0",
    "konlpy>=0.6.0",
    "matplotlib>=3.10.7",
    "openai>=1.60.0",
    "pandas>=2.3.3",
    "plotly>=5.24.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
    "selenium>=4.38.0",
    "streamlit>=1.41.1",
    "streamlit-plotly-events>=0.0.6",
    "webdriver-manager>=4.0.2",
    "wordcloud>=1.9.4",
]
